+++ src/main/java/org/jabref/preferences/AiPreferences.java
@@ -1,7 +1,13 @@
 package org.jabref.preferences;
 
 import javafx.beans.property.BooleanProperty;
+import javafx.beans.property.DoubleProperty;
+import javafx.beans.property.FloatProperty;
+import javafx.beans.property.IntegerProperty;
 import javafx.beans.property.SimpleBooleanProperty;
+import javafx.beans.property.SimpleDoubleProperty;
+import javafx.beans.property.SimpleFloatProperty;
+import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 
@@ -9,9 +15,23 @@ public class AiPreferences {
     private final BooleanProperty enableChatWithFiles;
     private final StringProperty openAiToken;
 
-    public AiPreferences(boolean enableChatWithFiles, String openAiToken) {
+    private final StringProperty systemMessage;
+    private final IntegerProperty messageWindowSize;
+    private final IntegerProperty documentSplitterChunkSize;
+    private final IntegerProperty documentSplitterOverlapSize;
+    private final IntegerProperty ragMaxResultsCount;
+    private final DoubleProperty ragMinScore;
+
+    public AiPreferences(boolean enableChatWithFiles, String openAiToken, String systemMessage, int messageWindowSize, int documentSplitterChunkSize, int documentSplitterOverlapSize, int ragMaxResultsCount, double ragMinScore) {
         this.enableChatWithFiles = new SimpleBooleanProperty(enableChatWithFiles);
         this.openAiToken = new SimpleStringProperty(openAiToken);
+
+        this.systemMessage = new SimpleStringProperty(systemMessage);
+        this.messageWindowSize = new SimpleIntegerProperty(messageWindowSize);
+        this.documentSplitterChunkSize = new SimpleIntegerProperty(documentSplitterChunkSize);
+        this.documentSplitterOverlapSize = new SimpleIntegerProperty(documentSplitterOverlapSize);
+        this.ragMaxResultsCount = new SimpleIntegerProperty(ragMaxResultsCount);
+        this.ragMinScore = new SimpleDoubleProperty(ragMinScore);
     }
 
     public BooleanProperty enableChatWithFilesProperty() {
@@ -36,5 +56,77 @@ public class AiPreferences {
 
     public void setOpenAiToken(String openAiToken) {
         this.openAiToken.set(openAiToken);
+    }
+
+    public StringProperty systemMessageProperty() {
+        return systemMessage;
+    }
+
+    public String getSystemMessage() {
+        return systemMessage.get();
+    }
+
+    public void setSystemMessage(String systemMessage) {
+        this.systemMessage.set(systemMessage);
+    }
+
+    public IntegerProperty messageWindowSizeProperty() {
+        return messageWindowSize;
+    }
+
+    public int getMessageWindowSize() {
+        return messageWindowSize.get();
+    }
+
+    public void setMessageWindowSize(int messageWindowSize) {
+        this.messageWindowSize.set(messageWindowSize);
+    }
+
+    public IntegerProperty documentSplitterChunkSizeProperty() {
+        return documentSplitterChunkSize;
+    }
+
+    public int getDocumentSplitterChunkSize() {
+        return documentSplitterChunkSize.get();
+    }
+
+    public void setDocumentSplitterChunkSize(int documentSplitterChunkSize) {
+        this.documentSplitterChunkSize.set(documentSplitterChunkSize);
+    }
+
+    public IntegerProperty documentSplitterOverlapSizeProperty() {
+        return documentSplitterOverlapSize;
+    }
+
+    public int getDocumentSplitterOverlapSize() {
+        return documentSplitterOverlapSize.get();
+    }
+
+    public void setDocumentSplitterOverlapSize(int documentSplitterOverlapSize) {
+        this.documentSplitterOverlapSize.set(documentSplitterOverlapSize);
+    }
+
+    public IntegerProperty ragMaxResultsCountProperty() {
+        return ragMaxResultsCount;
+    }
+
+    public int getRagMaxResultsCount() {
+        return ragMaxResultsCount.get();
+    }
+
+    public void setRagMaxResultsCount(int ragMaxResultsCount) {
+        this.ragMaxResultsCount.set(ragMaxResultsCount);
+    }
+
+    public DoubleProperty ragMinScoreProperty() {
+        return ragMinScore;
+    }
+
+    public double getRagMinScore() {
+        return ragMinScore.get();
+    }
+
+    public void setRagMinScore(double ragMinScore) {
+        this.ragMinScore.set(ragMinScore);
     }
 }